code rationalization done
