+++ sshmenuc/__init__.py
@@ -1,1 +1,14 @@
-from .sshmenuc import ConnectionManager
+from .core import ConnectionManager, ConnectionNavigator, SSHLauncher
+from .ui import Colors, MenuDisplay
+from .utils import setup_logging, setup_argument_parser
+
+__version__ = "0.1.0"
+__all__ = [
+    'ConnectionManager', 
+    'ConnectionNavigator', 
+    'SSHLauncher',
+    'Colors', 
+    'MenuDisplay',
+    'setup_logging',
+    'setup_argument_parser'
+]
