+++ sshmenuc/core/__init__.py
@@ -1,0 +1,9 @@
+"""
+Core module per sshmenuc.
+"""
+from .base import BaseSSHMenuC
+from .config import ConnectionManager
+from .navigation import ConnectionNavigator
+from .launcher import SSHLauncher
+
+__all__ = ['BaseSSHMenuC', 'ConnectionManager', 'ConnectionNavigator', 'SSHLauncher']
