+++ sshmenuc/utils/__init__.py
@@ -1,0 +1,6 @@
+"""
+Utils module per sshmenuc.
+"""
+from .helpers import setup_argument_parser, setup_logging, get_default_config_path, validate_host_entry
+
+__all__ = ['setup_argument_parser', 'setup_logging', 'get_default_config_path', 'validate_host_entry']
